RUI5

+++ techpack/audio/asoc/codecs/aw882xx/aw_dsp.c
@@ -1223,8 +1223,9 @@ int aw_get_algo_version(struct aw_device *aw_dev, char *algo_ver_buf)
 			return ret;
 	} else {
 		algo_data = (char *)&algo_ver;
-		snprintf(algo_ver_buf, ALGO_VERSION_MAX, "aw_algo_v%d.%d.%d.%d",
-				algo_data[3], algo_data[2], algo_data[1], algo_data[0]);
+		snprintf(algo_ver_buf, ALGO_VERSION_MAX, "aw_algo_v%u.%u.%u.%u",
+				(uint32_t)algo_data[3], (uint32_t)algo_data[2],
+				(uint32_t)algo_data[1], (uint32_t)algo_data[0]);
 	}
 
 	aw_dev_dbg(aw_dev->dev, "%s", algo_ver_buf);
